+++ ui/src/providers.js
@@ -1,10 +1,10 @@
 import SelfHostedOptions from "@/components/settings/providers/SelfHostedOptions.svelte";
 import MicrosoftOptions  from "@/components/settings/providers/MicrosoftOptions.svelte";
-import AuthentikOptions  from "@/components/settings/providers/AuthentikOptions.svelte";
+import OIDCOptions  from "@/components/settings/providers/OIDCOptions.svelte";
 
 // Object list with all supported OAuth2 providers in the format:
 // ```
-// { settingsKey: { title, icon, optionsComponent? } }
+// { settingsKey: { title, icon, hidden, optionsComponent? } }
 // ```
 //
 // If `optionsComponent` is provided it will receive 2 parameters:
@@ -70,9 +70,21 @@ export default {
         title: "LiveChat",
         icon:  "ri-chat-1-fill",
     },
-    authentikAuth: {
-        title: "Authentik",
+    oidcAuth: {
+        title: "OpenID Connect (Authentik, Keycloak, Okta, ...)",
         icon:  "ri-lock-fill",
-        optionsComponent: AuthentikOptions,
+        optionsComponent: OIDCOptions,
+    },
+    oidc2Auth: {
+        title: "(2) OpenID Connect (Authentik, Keycloak, Okta, ...)",
+        icon:  "ri-lock-fill",
+        hidden: true,
+        optionsComponent: OIDCOptions,
+    },
+    oidc3Auth: {
+        title: "(3) OpenID Connect (Authentik, Keycloak, Okta, ...)",
+        icon:  "ri-lock-fill",
+        hidden: true,
+        optionsComponent: OIDCOptions,
     },
 };